#81 Creacion de Relacion Cooperativa-Unidades

+++ database/migrations/2020_07_08_120144_create_rutas_table.php
@@ -16,7 +16,7 @@ class CreateRutasTable extends Migration
         Schema::create('rutas', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name',50);
-            $table->decimal('units_number',3);
+            $table->integer('units_number');
             $table->timestamps();
         });
     }
